+++ libs/ajax-action/post-actions/src/ajax-post-action-close/ajax-post-action-close.service.ts
@@ -1,0 +1,17 @@
+import { Injectable, Injector } from '@angular/core';
+import { AjaxPostActionHandler } from '@spryker/ajax-action';
+import { DrawerRef } from '@spryker/drawer';
+
+import { AjaxPostActionClose } from './types';
+
+/**
+ * Closes the drawer via {@link DrawerRef}
+ */
+@Injectable({
+  providedIn: 'root',
+})
+export class AjaxPostActionCloseService implements AjaxPostActionHandler {
+  handleAction(action: AjaxPostActionClose, injector: Injector): void {
+    injector.get(DrawerRef).close();
+  }
+}