+++ libs/ajax-action/post-actions/src/index.ts
@@ -1,1 +1,2 @@
 export * from './ajax-post-action-redirect';
+export * from './ajax-post-action-close';