+++ libs/table/src/lib/table/table.component.ts
@@ -47,7 +47,6 @@ import { TableFeatureEventBus } from '../table-feature/table-feature-event-bus';
 import { TableFeatureComponent } from '../table-feature/table-feature.component';
 import { TableFeatureDirective } from '../table-feature/table-feature.directive';
 import { TableFeaturesRendererService } from '../table-features-renderer/table-features-renderer.service';
-import { TableActionService } from './action.service';
 import { ColTplDirective } from './col-tpl.directive';
 import { TableColumnsResolverService } from './columns-resolver.service';
 import { TableDataConfiguratorService } from './data-configurator.service';
@@ -78,9 +77,9 @@ const shareReplaySafe: <T>() => MonoTypeOperatorFunction<T> = () =>
   providers: [
     TableDataConfiguratorService,
     TableColumnsResolverService,
-    TableActionService,
     TableFeaturesRendererService,
     TableDatasourceService,
+    TableActionsService,
   ],
 })
 export class CoreTableComponent